requirements.txt
django==3.1
psycopg2-binary==2.8.5
pillow==7.1
djangorestframework==3.11
drf-yasg
python-decouple==3.4
